pyproject.toml
[tool.poetry]
name = "skycock-match-manager"
version = "0.1.0"
description = ""
authors = ["jlee3120 <lozente@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.2"
Flask-SQLAlchemy = "^3.1.1"
Flask-Migrate = "^4.0.5"
pre-commit = "^3.6.2"
SQLAlchemy-Utils = "^0.41.1"
phonenumbers = "^8.13.30"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
